src/requirements.txt

fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2
pydantic-settings==2.2.1
email-validator>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
jinja2>=3.1
passlib[bcrypt]>=1.7
PyJWT>=2.9
python-multipart>=0.0.9
python-dotenv>=1.0
pandas>=2.2
scikit-learn>=1.5
joblib>=1.4
pytest>=8.2
httpx>=0.27
aiofiles==23.2.1
langchain==0.1.20
PyMuPDF==1.24.3
motor==3.4.0